Serverless function

+++ src/main/java/com/trading/signal/exchange/BinanceData.java
@@ -48,7 +48,7 @@ public class BinanceData {
                     Float.parseFloat((String) candle.get(2)),
                     Float.parseFloat((String) candle.get(3)),
                     Float.parseFloat((String) candle.get(4)),
-                    Integer.parseInt((String) candle.get(5))
+                    Float.parseFloat((String) candle.get(5))
             );
 
         }
